+++ js/state.js
@@ -1,5 +1,12 @@
 const AppState={
- project:{title:'The Worst Birthday Ever',theme:'VihuPlanet'},
+ project:{
+   title:'The Worst Birthday Ever',
+   author:'',
+   bookTitle:'The Worst Birthday Ever',
+   theme:'default',
+   createdDate:'',
+   modifiedDate:''
+ },
  slides:[],
  currentSlide:0
 };